pyproject.toml
[project]
name = "template-fastapi"
version = "0.1.0"
description = "FastAPI Template with Clean Architecture and SOLID Principles"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.13.3",
    "asgi-correlation-id>=4.3.4",
    "fastapi>=0.115.5",
    "httpx>=0.27.2",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.6.0",
    "pyhumps>=3.8.0",
    "python-multipart>=0.0.12",
    "returns>=0.24.0",
    "sqlalchemy-utils>=0.41.2",
    "sqlmodel>=0.0.22",
    "structlog>=24.4.0",
    "uvicorn[standard]>=0.32.0",
]

[tool.uv]
dev-dependencies = [
    "faker>=30.6.0",
    "ipython>=8.28.0",
    "isort>=5.13.2",
    "mypy>=1.12.0",
    "pytest>=8.3.3",
    "pytest-cov>=5.0.0",
    "pytest-dotenv>=0.5.2",
    "pytest-mock>=3.14.0",
    "pytest-sugar>=1.0.0",
    "pytest-vcr>=1.0.2",
    "ruff>=0.7.0",
]

[tool.ruff]
line-length = 79
target-version = "py313"

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

preview = true

[tool.ruff.lint]
fixable = ["ALL"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "single"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"

[tool.pytest.ini_options]
env_files = [
    ".env",
    ".test.env"
]